refactor: User2Dto 수정
