+++ lib/data_source/user/user2_data_source.dart
@@ -1,5 +1,5 @@
 import 'package:modu_3_dart_study/core/response.dart';
-import 'package:modu_3_dart_study/dto/user/user2_dto.dart';
+import 'package:modu_3_dart_study/dto/user2/user2_dto.dart';
 
 abstract interface class User2DataSource {
   Future<Response<List<User2Dto>>> getUsers();
