+++ test/mocks/mock_user2_data_source_impl.dart
@@ -1,6 +1,6 @@
 import 'package:modu_3_dart_study/core/response.dart';
 import 'package:modu_3_dart_study/data_source/user/user2_data_source.dart';
-import 'package:modu_3_dart_study/dto/user/user2_dto.dart';
+import 'package:modu_3_dart_study/dto/user2/user2_dto.dart';
 
 class MockUser2DataSourceImpl implements User2DataSource {
   final jsons = [
